% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlity_functions.R
\name{count_na}
\alias{count_na}
\title{Function that counts the number of NAs in a data frame or a vector}
\usage{
count_na(x)
}
\arguments{
\item{x}{The data frame or vector that should have its NA values counted.}
}
\description{
Function that counts the number of NAs in a data frame or a vector
}
