% tinyrox says don't edit this manually, but it can't stop you!
\name{.caption_file}
\alias{.caption_file}
\title{First subtitle file referenced by any caption track}
\usage{
.caption_file(ctracks, media_dir = NULL)
}
\arguments{
\item{ctracks}{List of caption Tracks.}

\item{media_dir}{Base directory for relative urls, or NULL.}
}
\value{
A subtitle path, or NULL.
}
\description{
First subtitle file referenced by any caption track
}
\keyword{internal}
