% tinyrox says don't edit this manually, but it can't stop you!
\name{.framing_vf}
\alias{.framing_vf}
\title{Build the scale/pad ffmpeg filter chain from a framing list}
\usage{
.framing_vf(framing)
}
\arguments{
\item{framing}{A framing list, or NULL.}
}
\value{
Character vector of filter expressions (possibly empty).
}
\description{
Build the scale/pad ffmpeg filter chain from a framing list
}
\keyword{internal}
