% tinyrox says don't edit this manually, but it can't stop you!
\name{.is_image}
\alias{.is_image}
\title{Is this output path a still image?}
\usage{
.is_image(path)
}
\arguments{
\item{path}{Output path.}
}
\value{
TRUE for common still-image extensions.
}
\description{
Is this output path a still image?
}
\keyword{internal}
