% tinyrox says don't edit this manually, but it can't stop you!
\name{.motion_xy}
\alias{.motion_xy}
\title{Coerce a motion anchor to a length-2 numeric}
\usage{
.motion_xy(v, default)
}
\arguments{
\item{v}{Field value (vector or list of two numbers).}

\item{default}{Value when v is NULL.}
}
\value{
A numeric vector of length 2.
}
\description{
Coerce a motion anchor to a length-2 numeric
}
\keyword{internal}
