% tinyrox says don't edit this manually, but it can't stop you!
\name{.parse_timestamp}
\alias{.parse_timestamp}
\title{Parse HH:MM:SS.mmm timestamp to seconds}
\usage{
.parse_timestamp(x)
}
\arguments{
\item{x}{Character timestamp}
}
\value{
Numeric seconds
}
\description{
Parse HH:MM:SS.mmm timestamp to seconds
}
\keyword{internal}
