% tinyrox says don't edit this manually, but it can't stop you!
\name{.timeline_layout}
\alias{.timeline_layout}
\title{The timeline's cornball layout metadata, or NULL}
\usage{
.timeline_layout(timeline)
}
\arguments{
\item{timeline}{A rotio Timeline.}
}
\description{
The timeline's cornball layout metadata, or NULL
}
\keyword{internal}
