% tinyrox says don't edit this manually, but it can't stop you!
\name{.track_role}
\alias{.track_role}
\title{A track's cornball role ("" when untagged)}
\usage{
.track_role(track)
}
\arguments{
\item{track}{A rotio Track.}
}
\description{
A track's cornball role ("" when untagged)
}
\keyword{internal}
