% tinyrox says don't edit this manually, but it can't stop you!
\name{.track_span}
\alias{.track_span}
\title{The timeline length a track occupies, Gaps included}
\usage{
.track_span(seq_v, files)
}
\arguments{
\item{seq_v}{A \code{.video_sequence()} result.}

\item{files}{The pre-rendered files.}
}
\description{
The timeline length a track occupies, Gaps included
}
\keyword{internal}
