% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness2012.R
\name{fitness2012}
\alias{fitness2012}
\title{Apply Fitness (2012) algorithm to RCA data}
\usage{
fitness2012(df, N)
}
\arguments{
\item{df}{data frame that contains columns "region", "unit", "rca" (binary).}

\item{N}{number of iterations the algorithm should run over.}
}
\value{
dataframe that contains both unit complexity and region fitness.
}
\description{
Apply Fitness (2012) algorithm to RCA data
}
