% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness2018.R
\name{fitness2018}
\alias{fitness2018}
\title{Apply Fitness (2018) algorithm to RCA data}
\usage{
fitness2018(df, N, delta = 0)
}
\arguments{
\item{df}{data frame that contains columns "region", "unit", "rca" (binary).}

\item{N}{number of iterations the algorithm should run over.}

\item{delta}{delta value used in function. default is 0.}
}
\value{
dataframe that contains both unit complexity and region fitness.
}
\description{
Apply Fitness (2018) algorithm to RCA data
}
