% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_average_sim.R
\name{get_average_sim}
\alias{get_average_sim}
\title{Find average similarity value in region}
\usage{
get_average_sim(rca_mat, sim_mat, exclude_0s = FALSE)
}
\arguments{
\item{rca_mat}{binary rca matrix with region rows, units in columns}

\item{sim_mat}{similarity matrix with units on both dimensions.}

\item{exclude_0s}{logical If TRUE, each 0 in the binary RCA matrix
is converted to NAs. This means that units not present in a region
does not pull the average down. That is, it is not the average value,
but the average value of units presents in the region. This pulls
non-diverse regions up.}
}
\value{
matrix with regions in rows, units in columns. The elements are
the average similarity-value in the region to the given unit (column).
}
\description{
In the rca matrix, each region (row) has a number of
units (col) with the value 1. From these units, the function
finds the average value of similarity to each unit.
}
