% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_closest_sim.R
\name{get_closest_sim}
\alias{get_closest_sim}
\title{Find highest similarity value in region}
\usage{
get_closest_sim(rca_mat, sim_mat)
}
\arguments{
\item{rca_mat}{binary rca matrix with region rows, units in columns}

\item{sim_mat}{similarity matrix with units on both dimensions.}
}
\value{
matrix with regions in rows, units in columns. The elements are
the highest similarity-value in the region to the given unit (column).
}
\description{
In the rca matrix, each region (row) has a number of
units (col) with the value 1. From these units, the function
finds the highest value of similarity to each unit.
}
