% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rca.R
\name{get_rca}
\alias{get_rca}
\title{Creates an incidence matrix from a region-unit intensity matrix using RCA}
\usage{
get_rca(mat, binary = TRUE)
}
\arguments{
\item{mat}{Matrix that contains regions on rows, units on columns and
intensity value in elements.}

\item{binary}{logical TRUE returns a binarized RCA matrix, where
values of at least 1 = 1, values below 1 = 0.}
}
\value{
Matrix that contains the revealed comparative
of region (rows) in unit (columns).
}
\description{
Takes a matrix with regions in rows and units in columns, where elements are
intensity. This could be countries as regions, products as units and export
value as intensity. Calculates the revealed comparative advantage for each
region-unit pair. Returns the incidence matrix.
}
