% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_similarity.R
\name{get_similarity}
\alias{get_similarity}
\title{Creates similarity matrices based on an input matrix.}
\usage{
get_similarity(mat, method, diag0 = TRUE)
}
\arguments{
\item{mat}{Binary matrix that contains regions in rows, units in columns and
presence/incidence in elements.}

\item{method}{String that lists the method to calculate similarity. Options
are: "proximty", "association", or jaccard". Only takes one argument. For
more similarity-values, run function again.}

\item{diag0}{Logical. Decides whether diagonal in rel_mat should be set to 0.}
}
\value{
rel_mat Unit x unit matrix that contains pair-wise similarity
value is elements. The diagonal is set to 0.
}
\description{
Takes an input matrix with regions in rows and units in columns,
where elements are a binary incidence (fx RCA). Calculates the
one of three possible similarity-metrics: proximity, Association
Strength, or Jaccard coefficient. Returns a unit x unit matrix
with pairwise similarity values as elements.
}
