% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weighted_average_sim.R
\name{get_weighted_average_sim}
\alias{get_weighted_average_sim}
\title{Find average similarity value in region, weighted by RCA}
\usage{
get_weighted_average_sim(rca_mat, sim_mat)
}
\arguments{
\item{rca_mat}{non-binary rca matrix with region rows, units in columns}

\item{sim_mat}{similarity matrix with units on both dimensions.}
}
\value{
matrix with regions in rows, units in columns. The elements are
the weighted average similarity-value in the region (row) to the given unit (column).
}
\description{
In the rca matrix, each region (row) has a number of
units (col) with the value 1. From these units, the function
finds the average value of similarity to each unit, weighted by the
RCA value in the rca_mat.
}
