% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp_testing.R
\name{null_samples}
\alias{null_samples}
\title{Non-parametric means test}
\usage{
null_samples(df, t0_threshold, trials)
}
\arguments{
\item{df}{data frame from the rel_* set}

\item{t0_threshold}{the t0_threshold used to compute is_new}

\item{trials}{the number samples to be drawn in the monte carlo.}
}
\description{
Non-parametric means test
}
